--- Trie/__init__.py
@@ -1,1 +1,0 @@
-from .trie import Trie
